Module rust_algorithm_club::collections

Collections.

Structs

BloomFilter
A space efficient probablistic data structures offering an approximate containment test with only false positive error.

Deque
A double-ended queue (abbreviated to deque), for which elements can be added or remove from both back and front ends.

HashMap
A hash map implemented with separate chaining collision resolution strategy.

HashSet
A hash set implementation based on HashMap.

Queue
A queue-like data structure implement through std::vec::Vec.

SinglyLinkedList
A singly-linked list with owned nodes.

Stack
A stack-like data structure implemented through a Vec.