Struct rust_algorithm_club::collections::Deque

pub struct Deque<T> { /* fields omitted */ }

A double-ended queue (abbreviated to deque), for which elements can be added or remove from both back and front ends.

Underneath the hood, this Deque uses a contiguous memory block as a ring buffer to store values.

References:

Rust Standard Library: std::collections::VecDeque

Wikipedia: Circular buffer

Implementations

impl<T> Deque<T>

pub fn new() -> Self

Constructs a new, empty Deque<T>.

For convenience, the deque initially allocates a region of a single T.

pub fn push_front(&mut self, elem: T)

Prepends the given element value to the beginning of the container.

Parameters

elem - The element to prepend.

Complexity

Constant.

pub fn push_back(&mut self, elem: T)

Appends the given element value to the end of the container.

Parameters

elem - The element to append.

Complexity

Constant.

pub fn pop_front(&mut self) -> Option<T>

Removes and returns the first element of the container. If there are no elements in the container, return None.

Complexity

Constant.

pub fn pop_back(&mut self) -> Option<T>

Removes and returns the last element of the container. If there are no elements in the container, return None.

Complexity

Constant.

pub fn front(&self) -> Option<&T>

Peeks the first element of the container. If there are no elements in the container, return None.

Complexity

Constant.

pub fn back(&self) -> Option<&T>

Peeks the last element of the container. If there are no elements in the container, return None.

Complexity

Constant.

pub fn is_empty(&self) -> bool

Checks whether the container is empty.

Complexity

Constant.

pub fn len(&self) -> usize

Gets the number of elements in the container.

Complexity

Constant.

pub fn iter(&self) -> Iter<'_, T>

Creates an iterator that yields immutable reference of each element.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Creates an iterator that yields mutable reference of each element.