Struct rust_algorithm_club::collections::HashSet

pub struct HashSet<T> where
    T: Hash + Eq,  { /* fields omitted */ }

A hash set implementation based on HashMap

Methods

impl<T> HashSet<T> where
T: Hash + Eq,

pub fn new() -> Self

Create an empty set.

pub fn len(&self) -> usize

Get the number of non-repetitive items.

pub fn is_empty(&self) -> bool

Return whether there is no any item in the set.

pub fn insert(&mut self, value: T) -> bool

Try to insert an item. Returns true if there were no such item in the set, returns false if an identical item is already in the set.

pub fn contains<Q: ?Sized>(&self, value: &Q) -> bool where
T: Borrow<Q>,
Q: Hash + Eq,

Returns whether an item is present in the set.

pub fn remove<Q: ?Sized>(&mut self, value: &Q) -> bool where
T: Borrow<Q>,
Q: Hash + Eq,

Try to remove an item from the set. Returns true if such item was present and removed, returns false if no such item was found in the set.

pub fn iter(
&self
) -> impl Iterator<Item = &T>

Creates an iterator that yields immutable reference of each item in arbitrary order.

pub fn union<'a>(
&'a self,
other: &'a HashSet<T>
) -> impl Iterator<Item = &T> + 'a

Returns an iterator visiting every item(s) that exists in self, in other, or in both self and other

pub fn difference<'a>(
&'a self,
other: &'a HashSet<T>
) -> impl Iterator<Item = &T>

Returns an iterator visiting every item(s) that exists in self but not in other.

pub fn symmetric_difference<'a>(
&'a self,
other: &'a HashSet<T>
) -> impl Iterator<Item = &T>

Returns an iterator visiting every item(s) that only exists in either self or other.

pub fn intersection<'a>(
&'a self,
other: &'a HashSet<T>
) -> impl Iterator<Item = &T>

Returns an iterator visiting every item(s) that exists in both self and other.