rust_algorithm_club::collections
Struct Queue

pub struct Queue<T> { /* private fields */ }

A queue-like data structure implement through std::vec::Vec.

This is a naive implementation whose insertion time complexity is O(n), which can be improved trivially by using a Deque or SinglyLinkedList.

References:

Queue (abstract data type)

Implementations

impl<T> Queue<T>

pub fn new() -> Self

Initialize a queue with empty vector

pub fn enqueue(&mut self, item: T)

Adds an element into queue.

Complexity

Constant.

pub fn dequeue(&mut self) -> Option<T>

Removes the oldest added element in queue.

Complexity

Linear in the size of the container.

pub fn peek(&self) -> Option<&T>

Retrieves the least recently added element without dequeuing.

Complexity

Constant.

pub fn size(&self) -> usize

Retrieves the size of the queue.

Complexity

Constant.