Struct rust_algorithm_club::collections::SinglyLinkedList

pub struct SinglyLinkedList<T> { /* fields omitted */ }

A singly-linked list with owned nodes.

This implementation is a simplified version of std::forward_list in C++.

References:

Rust Standard Library: std::collections::LinkedList

Learning Rust With Entirely Too Many Linked Lists

Stack Overflow: Deleting a node from singly linked list …

Stack Exchange: Reversal of a singly-linked list in Rust

Implementations

impl<T> SinglyLinkedList<T>

pub fn new() -> Self

Constructs a new, empty SinglyLinkedList<T>.

The list will not allocate until elements are pushed onto it.

pub fn push_front(&mut self, elem: T)

Prepends the given element value to the beginning of the container.

Parameters

elem - The element to prepend.

Complexity

Constant.

pub fn pop_front(&mut self) -> Option<T>

Removes and returns the first element of the container. If there are no elements in the container, return None.

Complexity

Constant.

pub fn insert_after(&mut self, pos: usize, elem: T) -> Result<(), usize>

Inserts an element after the specified position in the container.

If the position is out of bound, returns an Result:Err with the size of the list.

Parameters

pos - The index after which the element will be inserted.

elem - The element to be inserted.

Complexity

Search time O(n) + O(1).

pub fn remove(&mut self, pos: usize) -> Option<T>

Removes and returns an element at specified position from the container.

Parameters

pos - The index at which the element will be moved.

Complexity

Search time O(n) + constant.

pub fn clear(&mut self)

Removes all elements from the container.

Complexity

Linear in the size of the container.

pub fn is_empty(&self) -> bool

Checks whether the container is empty.

Complexity

Constant.

pub fn len(&self) -> usize

Gets the number of elements in the container.

Complexity

Linear in the size of the container.

pub fn reverse(&mut self)

Reverses the order of the elements in the container.

Complexity

Linear in the size of the container.

pub fn iter(&self) -> Iter<'_, T>

Creates an iterator that yields immutable reference of each element.

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Creates an iterator that yields mutable reference of each element.