rust_algorithm_club::collections
Struct Stack

pub struct Stack<T> { /* private fields */ }

A stack-like data structure implemented through a Vec.

The name “stack” for this type of structure comes from the analogy to a set of physical items stacked on top of each other, which makes it easy to take an item off the top of the stack, while getting to an item deeper in the stack may require taking off multiple other items first.

Considered as a linear data structure, or more abstractly a sequential collection, the push and pop operations occur only at one end of the structure, referred to as the top of the stack.

References:

Stack (abstract data type)

Big-O Algorithm Complexity Cheat Sheet

Implementations

impl<T> Stack<T>

pub fn with_capacity(maxsize: usize) -> Self

Initialize a stack of certain capacity.

Parameters

maxsize: Capacity of the collection. It limits how many items can be stored.

pub fn pop(&mut self) -> Option<T>

Removes the most recently added element that was not yet removed.

Returns

Returns the most recently added item. If nothing was added, None will be returned.

Complexity

Constant.

pub fn push(&mut self, item: T) -> bool

Adds an element to the collection.

Returns

Returns true if the collection has space left and item is successfully added, otherwise returns false.

Complexity

Constant.

pub fn size(&self) -> usize

Returns

Returns the size of collection, indicates how many items are added in the collection.

Note

Size and capacity are different concepts. Capacity limits how many items can be stored, while size indicates how many items is currently stored.

pub fn peek(&self) -> Option<&T>

Peeks the last element added without tampering the collection.

Returns

Returns the most recently added item. If nothing was added, None will be returned.