Module rust_algorithm_club::searching

Searching algorithms.

Functions

binary_search

Handmade binary search for a sorted sequence.

exponential_search

Exponential search. An binary search variant that can perform on unbounded sequences and infinite lists.

interpolation_search

Search in sorted sequences by checking the next position based on an linear interpolation of the search key.

linear_search

Linear search.