Crate rust_algorithm_club

Welcome to Rust algorithm club!

Modules

collections

Colletions are general-purpose or specialized data structures in which elements store.

searching

Searching algorithms.

sorting

Sorting algorithms.

Functions

hamming_distance

Calculate Hamming distance to two unsigned intergers.

hamming_distance_str

Calculate Hamming distance of two UTF-8 encoded strings.

levenshtein_distance

Calculate Levenshtein distance for two UTF-8 encoded strings.

levenshtein_distance_naive

Naïvely calculate Levenshtein distance using the whole distance matrix to store information for all substrings.