Struct rust_algorithm_club::collections::HashMap

pub struct HashMap<K, V> where
    K: Hash + Eq,  { /* fields omitted */ }

A hash map implemented with separate chaining collision resolution strategy.

This implementation is focused on hash map funcionalities, so we choose to adopt Rust DefaultHasher to avoid avalanche of details, and vectors as the underlying data structure for separate chaining method.

The interface is a simplified version of Rust HashMap.

Reference:

Rust Standard Library: std::collections::HashMap

C++ Container Library: std::unordered_map

Methods

impl<K, V> HashMap<K, V> where
K: Hash + Eq,

pub fn new() -> Self

Creates an empty map with capacity 0.

The allocation is triggered at the first insertion occurs.

pub fn with_capacity(cap: usize) -> Self

Creates a map with a given capacity as the number of underlying buckets.

Parameters

cap: The number of bucket in the map.

pub fn get(&self, key: &K) -> Option<&V>

Gets a reference to the value under the specified key.

TODO: To treat owned and borrowed values in equivalent ways as other collections in std do, we should use Borrow trait to abstract over the type of key to hash. This concept can also applied for get_mut remove, and other operations that constrain by the type system..

Some useful resources:

Trait std::borrow:Borrow

TRPL 1st edition: Brrrow and AsRef

Complexity

Constant (amortized).

pub fn get_mut(&mut self, key: &K) -> Option<&mut V>

Gets a mutable reference to the value under the specified key.

Complexity

Constant (amortized).

pub fn insert(&mut self, key: K, value: V) -> Option<V>

Inserts key-value pair into the map. Replaces previous value if the same key exists at the same index.

Returns the old value if the key presents. Otherwise returnes None.

Steps are described as following:

Try to resize hashmap to ensure an appropriate load factor.

Compute hash of the key to get the inner bucket under certain index.

Find if there is already a pair with identical key.

If yes, substitue for it.

Else, push new value into the bucket.

Parameters

key - Key of the pair to insert.

value - Value of the pair to insert.

Complexity

Constant (amortized).

pub fn remove(&mut self, key: &K) -> Option<V>

Removes a pair with specified key.

The caveat is that ordering in the bucket cannot be preserved due to the removal using swap_remove to ensure O(1) deletion.

Parameters

key - Key of the pair to remove.

Complexity

Constant. This operation won't shrink to fit automatically.

pub fn clear(&mut self)

Removes all key-value pairs but keeps the allocated memory for reuse.

Complexity

Linear in the size of the container.

pub fn is_empty(&self) -> bool

Checks whether the container is empty.

Complexity

Constant.

pub fn len(&self) -> usize

Gets the number of key-value pairs in the container.

Complexity

Constant.

pub fn bucket_count(&self) -> usize

Gets the number of underlying buckets.

Complexity

Constant.