Struct rust_algorithm_club::collections::HashSet

pub struct HashSet<T> where
    T: Hash + Eq,  { /* fields omitted */ }

A hash set implementation based on HashMap.

References:

Rust Standard Library: std::collections::HashSet

Methods

impl<T> HashSet<T> where
T: Hash + Eq,

pub fn new() -> Self

Creates an empty set.

pub fn len(&self) -> usize

Gets the number of non-repetitive elements, equivalently to the cardinality of a set.

Complexity

Constant.

pub fn is_empty(&self) -> bool

Returns whether there is no any element in the set.

Complexity

Constant.

pub fn insert(&mut self, value: T) -> bool

Inserts an element into the set.

Returns true if there were no such element in the set; returns false if an identical element is already in the set.

Parameters

value - Element to be inserted.

Complexity

Constant.

pub fn contains<Q: ?Sized>(&self, value: &Q) -> bool where
T: Borrow<Q>,
Q: Hash + Eq,

Returns whether an element is present in the set.

This is equivalent to "belongs to ∈" relation in mathematics.

Parameters

value - Element to be checked whether is in the set.

Complexity

Constant.

pub fn remove<Q: ?Sized>(&mut self, value: &Q) -> bool where
T: Borrow<Q>,
Q: Hash + Eq,

Removes an element from the set.

Returns true if such item was present and removed; returns false if no such item was found in the set.

Parameters

value - Element to be removed.

Complexity

Constant.

pub fn iter(
&self
) -> impl Iterator<Item = &T>

Creates an iterator yielding immutable reference of each item in arbitrary order.

pub fn union<'a>(
&'a self,
other: &'a HashSet<T>
) -> impl Iterator<Item = &T>

Returns an iterator visiting items that exists in self, in other, or in both self and other

This is equivalent to self ∪ other in mathematics.

Parameters

other - The other set.

pub fn difference<'a>(
&'a self,
other: &'a HashSet<T>
) -> impl Iterator<Item = &T>

Returns an iterator visiting items that exists in self but not in other.

This is equivalent to self \ other in mathematics.

Parameters

other - The other set.

pub fn symmetric_difference<'a>(
&'a self,
other: &'a HashSet<T>
) -> impl Iterator<Item = &T>

Returns an iterator visiting items that only exists in either self or other but not in their intersection.

This is equivalent to self △ other in mathematics.

Parameters

other - The other set.

pub fn intersection<'a>(
&'a self,
other: &'a HashSet<T>
) -> impl Iterator<Item = &T>

Returns an iterator visiting items that exists in both self and other.

This is equivalent to self ∩ other in mathematics.

Parameters

other - The other set.

pub fn is_disjoint(&self, other: &HashSet<T>) -> bool

Returns true if self has no elements in common with other.

This is equivalent to checking for an empty intersection, which means their intersection is the empty set ∅.

Parameters

other - The other set.

Complexity

Linear in the size of self.

pub fn is_subset(&self, other: &HashSet<T>) -> bool

Returns true if other contains at least all elements in self.

This is equivalent to self ⊆ other in mathematics.

Parameters

other - The other set.

Complexity

Linear in the size of self.

pub fn is_superset(&self, other: &HashSet<T>) -> bool

Returns true if self contains at least all elements in other.

This is equivalent to self ⊇ other in mathematics.

Parameters

other - The other set.

Complexity

Linear in the size of other.